# knowsphere-ai-complete_timestamp/knowsphere-ai/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: KnowSphere AI API
  version: "6.0.0"
  description: |
    Enterprise Knowledge Intelligence Platform — REST API.

    Scoping note: this specification covers the primary request/response
    shape of every resource group (auth, providers, documents, chat,
    analytics, audit, observability, notifications, health). It documents
    the core, most-used parameters for each endpoint rather than every
    optional query parameter combination across 60+ endpoints — treat this
    as the authoritative contract for integration purposes, and the actual
    route source (`app/*/routes.py`) as the final word on any parameter
    not shown here.

    All endpoints are prefixed with `/api/v1`. Authentication is a JWT
    Bearer token in the `Authorization` header, obtained via `/auth/login`.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        file_type: { type: string, example: pdf }
        source_type: { type: string, enum: [upload, email, chat_export, share_link] }
        status: { type: string, enum: [uploaded, validating, parsing, chunking, embedding, indexing, ready, failed] }
        department: { type: string, nullable: true }
        chunk_count: { type: integer }
        visible_to_roles: { type: array, items: { type: string } }
...
paths:
...
  /documents:
    get:
      summary: List documents (ACL-filtered for non-managers)
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: file_type
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items: { $ref: "#/components/schemas/Document" }
    post:
      summary: Upload one or more documents (admin/manager only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items: { type: string, format: binary }
                department: { type: string }
      responses:
        "202": { description: "Upload accepted, processing asynchronously" }
...
  /analytics/trends:
    get:
      summary: Time-series activity trend
      parameters:
        - name: granularity
          in: query
          schema: { type: string, enum: [day, week, month] }
      responses:
        "200": { description: OK }
...
  /audit:
    get:
      summary: Search audit logs (admin only)
      parameters:
        - name: action
          in: query
          schema: { type: string }
        - name: actor_email
          in: query
          schema: { type: string }
      responses:
        "200": { description: OK }
  /audit/export:
    get:
      summary: Export audit logs (admin only)
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [csv, excel, pdf] }
      responses:
        "200": { description: File download }
...